pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pycodedownload"
version = "0.1.0"
description = "A package to download VSCode extensions from the Marketplace"
readme = "README.md"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["vscode", "extensions", "marketplace", "download"]
dependencies = [
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=12.0.0",
    "semver",
]
requires-python = ">=3.7"

[project.urls]
Homepage = "https://github.com/yourusername/pycodedownload"

[project.scripts]
pycodedownload = "pycodedownload.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/pycodedownload"]
